[feat] Project overhaul - Changed and fixed a lot of things that are relatively old with the current standing of the project. - Check the PROJECT_STATUS.md for updates regarding the above.
Current status of the project:
- login and registration working with need to update the darkmode compatability.
- registration needs to be updated to work with the new email verification system.
- tests aren't super focused on currently so CI/CD will likely fail.
- everything is in developement mode so expect some bugs and some hardcoded URLS
- everything regarding prod is kind of neglected at the moment to focus on functionality.
- games import working (at least for chess.com, didn't check lichess)
- some components are kind of left behind currently, but will be updated in the future.
- the MAIN focus now is single game analysis and batch analysis.
- now I am still working on single game analysis. Due to reimplementing the whole analysis pipeline, the analysis pipeline is kind of messed up.
- the current status of single game analysis is the following:
the backend does recieve the game and celery creates a task for it and the task gets proceed (not purely in the background, but it's a simple modification so don't worry about that)
the frontend sends polling (status) requests to check the progress of the analysis (and updates the loading screen progress bar based on that).
The status requests for some reason always return "progress 0" not sure why, but I assume it's because of the status naming in the backend and the frontend.

Interesting bug: there is a dependcy for celery that is incompatible with python >= 3.12, the fix is to make a venv with python 3.11 and install the dependencies there.

To run the project, run the following command:
cd chess_mate
run.bat (if you are on mac, just convert the bat file to a sh file and run it, I was too lazy to do that)

The above should open three terminals, a django terminal, a redis terminal and a celery terminal.
The django terminal is the one that runs the server, the redis terminal is the one that runs the redis server and the celery terminal is the one that runs the celery worker.

You should monitor celery and django for errors, if you see any errors, please report them to me.

for the frontend, just run the following command:
cd chessmate/frontend
npm start

I know there are a couple of ESLint issues, but we can just ignore for now.

if you have any questions, just contact me.